refactor: chat service interface 생성
